Eliminate worker trait & move worker config into core
