+++ crates/sdk-core/src/core_tests/replay_flag.rs
@@ -3,17 +3,14 @@ use crate::{
     worker::{LEGACY_QUERY_ID, client::mocks::mock_worker_client},
 };
 use std::{collections::VecDeque, time::Duration};
-use temporalio_common::{
-    Worker as CoreWorker,
-    protos::{
-        TestHistoryBuilder, canned_histories,
-        coresdk::{
-            workflow_activation::{WorkflowActivationJob, workflow_activation_job},
-            workflow_completion::WorkflowActivationCompletion,
-        },
-        temporal::api::{enums::v1::EventType, query::v1::WorkflowQuery},
-        test_utils::{query_ok, start_timer_cmd},
+use temporalio_common::protos::{
+    TestHistoryBuilder, canned_histories,
+    coresdk::{
+        workflow_activation::{WorkflowActivationJob, workflow_activation_job},
+        workflow_completion::WorkflowActivationCompletion,
     },
+    temporal::api::{enums::v1::EventType, query::v1::WorkflowQuery},
+    test_utils::{query_ok, start_timer_cmd},
 };
 
 #[tokio::test]
