+++ crates/sdk-core/src/pollers/mod.rs
@@ -10,15 +10,13 @@ pub use temporalio_client::{
 use crate::{
     abstractions::{OwnedMeteredSemPermit, TrackedOwnedMeteredSemPermit},
     telemetry::metrics::MetricsContext,
+    worker::{ActivitySlotKind, NexusSlotKind, SlotKind, WorkflowSlotKind},
 };
 use anyhow::{anyhow, bail};
 use futures_util::{Stream, stream};
 use std::{fmt::Debug, marker::PhantomData};
-use temporalio_common::{
-    protos::temporal::api::workflowservice::v1::{
-        PollActivityTaskQueueResponse, PollNexusTaskQueueResponse, PollWorkflowTaskQueueResponse,
-    },
-    worker::{ActivitySlotKind, NexusSlotKind, SlotKind, WorkflowSlotKind},
+use temporalio_common::protos::temporal::api::workflowservice::v1::{
+    PollActivityTaskQueueResponse, PollNexusTaskQueueResponse, PollWorkflowTaskQueueResponse,
 };
 use tokio::select;
 use tokio_util::sync::CancellationToken;
